Simplify implementation

+++ kpops/component_handlers/kubernetes/pvc_handler.py
@@ -17,35 +17,28 @@ class PVCHandler:
         config = KubeConfig.from_env()
         self._client = AsyncClient(config=config)
 
-    async def list_pvcs(self) -> list[str]:
-        async def read_pvcs() -> AsyncIterator[str]:
-            async for pvc in self._client.list(
-                PersistentVolumeClaim,
-                namespace=self.namespace,
-                labels={"app": self.app_name},
-            ):
-                if not pvc.metadata or not pvc.metadata.name:
-                    continue
-                yield pvc.metadata.name
+    async def list_pvcs(self) -> AsyncIterator[str]:
+        async for pvc in self._client.list(
+            PersistentVolumeClaim,
+            namespace=self.namespace,
+            labels={"app": self.app_name},
+        ):
+            if not pvc.metadata or not pvc.metadata.name:
+                continue
+            yield pvc.metadata.name
 
-        pvc_names = [pvc async for pvc in read_pvcs()]
-
+    async def delete_pvcs(self, dry_run: bool) -> None:
+        pvc_names = [pvc_name async for pvc_name in self.list_pvcs()]
         if not pvc_names:
             log.warning(
                 f"No PVCs found for app '{self.app_name}', in namespace '{self.namespace}'"
             )
-        else:
-            log.debug(
-                f"In namespace '{self.namespace}' StatefulSet '{self.app_name}' has corresponding PVCs: '{pvc_names}'"
-            )
-        return pvc_names
-
-    async def delete_pvcs(self) -> None:
-        pvc_names = await self.list_pvcs()
-
+            return
         log.debug(
             f"Deleting in namespace '{self.namespace}' StatefulSet '{self.app_name}' PVCs '{pvc_names}'"
         )
+        if dry_run:
+            return
         for pvc in pvc_names:
             await self._client.delete(
                 PersistentVolumeClaim, pvc, namespace=self.namespace
